Rename modules to continual_transformers
